Separate functionality into separate files
